json not wprking properly
